Add Skinduino demo app and overwrite previous version
