--- AnotherDemoTest/.gitignore
@@ -1,8 +1,0 @@
-*.iml
-.gradle
-/local.properties
-/.idea/workspace.xml
-/.idea/libraries
-.DS_Store
-/build
-/captures
