--- AnotherDemoTest/app/src/main/res/menu/main_menu.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
-    <item
-        android:id="@+id/menu_start"
-        android:title="START" />
-
-    <item
-        android:id="@+id/menu_stop"
-        android:title="STOP" />
-
-    <item
-        android:id="@+id/menu_test"
-        android:title="TEST" />
-</menu>
