--- AnotherDemoTest/app/src/main/res/values/strings.xml
@@ -1,7 +1,0 @@
-<resources>
-    <string name="app_name">AnotherDemoTest</string>
-
-    <string name="title_activity_image">ImageActivity</string>
-    <string name="dummy_button">Dummy Button</string>
-    <string name="dummy_content">DUMMY\nCONTENT</string>
-</resources>
