--- AnotherDemoTest/settings.gradle
@@ -1,1 +1,0 @@
-include ':app'
