+++ android/SkinduinoDemo/app/src/main/res/mipmap-xxhdpi/ic_launcher.png
