+++ android/SkinduinoDemo/app/src/main/res/mipmap-xxxhdpi/ic_launcher.png
