+++ android/SkinduinoDemo/app/src/main/res/values/dimens.xml
@@ -1,0 +1,6 @@
+<resources>
+    <!-- Default screen margins, per the Android Design guidelines. -->
+    <dimen name="activity_horizontal_margin">16dp</dimen>
+    <dimen name="activity_vertical_margin">16dp</dimen>
+    <dimen name="cell_size">90dp</dimen>
+</resources>
